pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conclave"
version = "0.1.0"
description = "Multi-agent meeting system where AI agents represent humans"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "litellm>=1.40",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
